dockerized the backend
