+++ docker-compose.yml
@@ -1,0 +1,24 @@
+services:
+  nigeria-constitution:
+    build: .
+    container_name: backend
+    ports:
+      - "8000:8000"
+    volumes:
+      - .:/app
+    command: uvicorn request:app --host 0.0.0.0 --port 8000 --reload
+    environment:
+      - OLLAMA_URL=http://ollama:11434 
+    depends_on:
+      - ollama
+
+  ollama:
+    image: ollama/ollama:latest
+    container_name: ollama
+    ports:
+      - "11434:11434"
+    volumes:
+    - ollama_models:/root/.ollama/models
+
+volumes:
+  ollama_models:
