+++ run.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+set -e
+docker-compose up --build
