+++ server/requirements.txt
@@ -1,0 +1,9 @@
+fastapi
+uvicorn
+pydantic
+langchain
+langchain-community
+langchain-chroma
+langchain-huggingface
+sentence-transformers
+langchain-ollama
